# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.CreateUserRequest:
    properties:
      email:
        example: example@gmail.com
        type: string
      first_name:
        example: John
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        example: Snow
        maxLength: 50
        minLength: 2
        type: string
      password:
        example: password
        maxLength: 32
        minLength: 6
        type: string
      username:
        example: username
        maxLength: 50
        minLength: 5
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  api.DefaultResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  api.GetUserByUsernameResponse:
    properties:
      body:
        $ref: '#/definitions/entity.User'
      code:
        type: integer
      message:
        type: string
    type: object
  api.GrantRoleToUser:
    properties:
      role:
        description: 'Allowed values: "user", "moderator", "admin"'
        example: ADMIN
        type: string
    required:
    - role
    type: object
  api.LoginRequest:
    properties:
      credentials:
        description: Can be username or email
        example: username
        type: string
      password:
        example: password
        type: string
    required:
    - credentials
    - password
    type: object
  api.LoginResponse:
    properties:
      body:
        $ref: '#/definitions/entity.Token'
      code:
        type: integer
      message:
        type: string
    type: object
...
  entity.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - USER
    - MODERATOR
    - ADMIN
...
  entity.Token:
    properties:
      expiry:
        type: string
      token:
        type: string
    type: object
  entity.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: dragon090986@gmail.com
    name: Kirill Shaforostov
  description: API for Library application
  title: Library API
  version: 0.0.1
paths:
...
  /mod/roles/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.GrantRoleToUser'
      produces:
      - application/json
      responses:
        "200":
          description: role of user was successfully updated
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update role of user by his ID. Requires ADMIN role
      tags:
      - Moderation
...
  /users/{username}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Username of user
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/api.GetUserByUsernameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get user by his username
      tags:
      - Users
  /users/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: User succesfully deleted
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Token succesfully created
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Authenticated user and return access token
      tags:
      - Authentication
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User succesfully created
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Register new user
      tags:
      - Users
...
  /users/update:
    patch:
      parameters:
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User succesfully updated
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user info
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"